skills/control_reaction_perceptor/pyproject.toml
[project]
name = "Control Reaction - Perceptor"
version = "0.1.0"
description = ""
authors = [{ name = "John Doe", email = "john.doe@composabl.com" }]
dependencies = [
    "composabl-core",
    "numpy"
]

[composabl]
type = "skill-teacher"
entrypoint = "control_reaction_perceptor.teacher:BaseCSTR"
